requirements.txt
aniso8601==10.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==6.0.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cssselect==1.3.0
cssutils==2.11.1
docx==0.2.4
et_xmlfile==2.0.0
flasgger==0.9.7.1
Flask==3.1.0
flask-cors==5.0.1
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
MarkupSafe==3.0.2
mistune==3.1.3
more-itertools==10.7.0
MouseInfo==0.1.3
numpy==2.2.4
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.1.0
premailer==3.10.0
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
soundfile==0.13.1
soupsieve==2.7
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
yagmail==0.15.293
gunicorn==21.2.0
